requirements.txt
# Core dependencies
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0

# Redis (for background task coordination)
redis[hiredis]>=5.0.0

# MongoDB (for event storage)
motor>=3.3.0
pymongo>=4.6.0

# PostgreSQL direct access
asyncpg>=0.29

# Socket.IO client (AI worker connection to socket-server)
python-socketio[asyncio_client]>=5.11

# LangChain and LangGraph
langchain-core==1.2.5
langchain-openai>=0.3.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.2

# Serialization
orjson==3.11.5
ormsgpack==1.12.1

# Data validation
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5

# Configuration
python-dotenv>=1.0.0
PyYAML==6.0.3

# HTTP
requests==2.32.5
requests-toolbelt==1.0.0

# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Utilities
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
xxhash==3.6.0
zstandard==0.25.0

# Excel
openpyxl>=3.1.0
